# WIP: Consolidate forward/reverse edge traversal in isochrone/reach code.