--- a/bench/thor/CMakeLists.txt
+++ b/bench/thor/CMakeLists.txt
@@ -1,3 +1,4 @@
 add_valhalla_benchmark(costmatrix)
 add_valhalla_benchmark(routes)
 add_valhalla_benchmark(isochrone)
+add_valhalla_benchmark(reach)