--- a/valhalla/thor/dijkstras.h
+++ b/valhalla/thor/dijkstras.h
@@ -188,7 +188,7 @@ protected:
   template <const ExpansionType expansion_direction>
   void ExpandInner(baldr::GraphReader& graphreader,
                    const baldr::GraphId& node,
-                   const decltype(Dijkstras::bdedgelabels_)::value_type& pred,
+                   const typename decltype(Dijkstras::bdedgelabels_)::value_type& pred,
                    const uint32_t pred_idx,
                    const baldr::DirectedEdge* opp_pred_edge,
                    const bool from_transition,