# Issue

This PR is the first of what I expect to be a series that work towards addressing https://github.com/valhalla/valhalla/issues/2885

There's lots of work to do, so in order to mimize conflict with other concurrent work, I plan to push early and often, and get these refactors into `master` iteratively, rather than in one big refactor nuclear explosion.

This PR addresses some of the code duplication within `dijkstra.cc` - in particular, the `Compute/ComputeReverse` functions.  Specifically, here's what's happened;

  - `Dijkstra::ExpandForward/ExpandReverse` have been consolidated into a single template function called `ExpandTemplate<direction>` whose body lives in `dijkstra.h` - there are `if` statements within that to handle the slight differences in forward/reverse expansion - final naming of this function is TBD, I just picked one to make it obvious.
  - Same for `Compute/ComputeReverse` - consolidated into `Compute<direction>` in `dijkstra.h`
  - All callers of these functions have been updated to call the correct templated version

I did *not* touch the `ComputeMultiModal` function yet - there's a lot more diff in there than between the forward/reverse expansion code.  I'd like to do it as a separate PR once we've settled this one.

This is a fairly small refactor - I expect the first few of these will trigger numerous discussions, and as we sort through the concerns, I'll be able to move through the remaining code more quickly.  With that in mind, I'm going to try to keep the PRs small and contained so we don't have to review/discuss huge sweeping sets of changes.

As I go through the code, I'm absolutely sure I'm going to find inconsistencies - that's the whole reason for undergoing this refactor in the first place.  I'm going to do my best to properly merge duplicated code and incorporate features that are implemented in one place but not in another, but when I'm unsure, I will put explicit questions/comments in the PR to ensure that the correct behaviour is implemented.

The overall intent here is pure refactoring.  I expect that some bugfixes that exist in one routing algorithm will become available in another routing algorithm that was stale, but other than that, my goal is to have no impact on performance or functionality with these changes.  As folks review these PRs, if you notice things that need to be fixed, let's make issues for them and keep these PRs focused on the refactoring process.

How did I do this?

I moved the functions that look like they had a lot of shared code into separate files, then diffed them.  I carefully reviewed the diffs, change-by-change, and created new consolidated versions of the functions with compile-time branches instead of separate mostly-duplicated copies.  Here's a sample diff of the `Expand/ExpandReverse` functions that got merged here:

<img width="1285" alt="Screen Shot 2021-03-25 at 6 42 31 PM" src="https://user-images.githubusercontent.com/1892250/112565330-1616b900-8d9a-11eb-8dfa-bd895658af70.png">

The actual functional differences are fairly small - some flipped variables in the reverse case, and some unnecessary different naming, and the odd special-case logic, mostly for the reverse case.

## Tasklist

 - [ ] Add tests
 - [ ] Add #fixes with the issue number that this PR addresses
 - [ ] Update the docs with any new request parameters or changes to behavior described
 - [ ] Update the [changelog](CHANGELOG.md)

## Requirements / Relations

 Link any requirements here. Other pull requests this PR is based on?