@kevinkreiser 2021-03-29
got to love that diff ratio:
![image](https://user-images.githubusercontent.com/697548/112886913-8d04c800-90a0-11eb-908c-1d9dd824f2c9.png)

so far this looks good to me. once CI is good with everything. i think all we have to do is run RAD and perf tests on it to confirm its within measurement error. since its mainly isochrones it can be a little bit less straightforward but there are modes on the `run_with_server.py` that will allow you to do both checking for diffs in results as well as a general performance comparison. just let me know if that makes sense as a way to confirm that the changes are good