@danpat 2021-03-29
@kevinkreiser 👍 I've got a `bench/thor/isochrone.cc` in the works to check the core algorithm - I'm not expecting any impact, most of the new branches are evaluated at compile time and disappear.  RAD tests TBD but I agree they should be done.