@danpat 2021-03-29
Benchmark results using a full planet extract.  TL;DR - looks like any differences are under the noise floor, I'd say "no significant impact on performance".

### Before 17e9a3dffb0c1d42dee695698527e9edec2a8f85
```
------------------------------------------------------------------------------------
Benchmark                                          Time             CPU   Iterations
------------------------------------------------------------------------------------
BM_IsochroneUtrecht/1/repeats:10_mean          0.957 ms        0.956 ms           10
BM_IsochroneUtrecht/2/repeats:10_mean           1.22 ms         1.22 ms           10
BM_IsochroneUtrecht/4/repeats:10_mean           1.73 ms         1.73 ms           10
BM_IsochroneUtrecht/8/repeats:10_mean           3.37 ms         3.37 ms           10
BM_IsochroneUtrecht/16/repeats:10_mean          9.97 ms         9.96 ms           10
BM_IsochroneUtrecht/32/repeats:10_mean           888 ms          887 ms           10
BM_IsochroneUtrecht/64/repeats:10_mean          2289 ms         2288 ms           10
BM_IsochroneUtrecht/120/repeats:10_mean         6046 ms         6043 ms           10
```

### After 17e9a3dffb0c1d42dee695698527e9edec2a8f85
```
------------------------------------------------------------------------------------
Benchmark                                          Time             CPU   Iterations
------------------------------------------------------------------------------------
BM_IsochroneUtrecht/1/repeats:10_mean           1.05 ms         1.05 ms           10
BM_IsochroneUtrecht/2/repeats:10_mean           1.30 ms         1.30 ms           10
BM_IsochroneUtrecht/4/repeats:10_mean           1.89 ms         1.89 ms           10
BM_IsochroneUtrecht/8/repeats:10_mean           3.58 ms         3.57 ms           10
BM_IsochroneUtrecht/16/repeats:10_mean          9.55 ms         9.54 ms           10
BM_IsochroneUtrecht/32/repeats:10_mean           868 ms          868 ms           10
BM_IsochroneUtrecht/64/repeats:10_mean          2293 ms         2292 ms           10
BM_IsochroneUtrecht/120/repeats:10_mean         6074 ms         6071 ms           10
```