@danpat 2021-03-29
Benchmark results for reachability (checks reachability of every edge in Utrecht):

### Before
```
BM_ReachUtrecht/repeats:10_mean         2713 ms         2696 ms           10
BM_ReachUtrecht/repeats:10_median       2663 ms         2661 ms           10
BM_ReachUtrecht/repeats:10_stddev        114 ms          103 ms           10
```

### After
```
BM_ReachUtrecht/repeats:10_mean         2696 ms         2695 ms           10
BM_ReachUtrecht/repeats:10_median       2689 ms         2688 ms           10
BM_ReachUtrecht/repeats:10_stddev       69.9 ms         69.2 ms           10
```