@kevinkreiser 2021-03-30
@danpat can you bench something that is significantly larger than Utrecht? this may be an unpopular opinion but that graph is so small most of the deviation has to be in the setup/teardown or can you modify the bench to be able to take a different dataset and run longer routes on that? or maybe you could clarify a bit about what you meant by full planet extract, i didnt quite understand that.