@kevinkreiser 2021-03-30
looks also like msvc doesnt like the syntax somewhere:

```
D:\a\1\s\src\thor\dijkstras.cc(172): error C2065: 'not': undeclared identifier
D:\a\1\s\src\thor\dijkstras.cc(282): note: see reference to function template instantiation 'void valhalla::thor::Dijkstras::ExpandInner<valhalla::thor::ExpansionType::forward>(valhalla::baldr::GraphReader &,const valhalla::baldr::GraphId &,const valhalla::sif::BDEdgeLabel &,const uint32_t,const valhalla::baldr::DirectedEdge *,const bool,const valhalla::baldr::TimeInfo &)' being compiled
D:\a\1\s\src\thor\dijkstras.cc(172): error C2146: syntax error: missing ')' before identifier 'FORWARD'
D:\a\1\s\src\thor\dijkstras.cc(172): error C2059: syntax error: ')'
D:\a\1\s\src\thor\dijkstras.cc(173): error C2059: syntax error: ';'
D:\a\1\s\src\thor\dijkstras.cc(189): error C2065: 'opp_edge': undeclared identifier
D:\a\1\s\src\thor\dijkstras.cc(189): error C2065: 't2': undeclared identifier
D:\a\1\s\src\thor\dijkstras.cc(189): error C2065: 'oppedgeid': undeclared identifier
D:\a\1\s\src\thor\dijkstras.cc(186): error C2789: 'allowed': an object of const-qualified type must be initialized
D:\a\1\s\src\thor\dijkstras.cc(186): note: see declaration of 'allowed'
D:\a\1\s\src\thor\dijkstras.cc(199): error C2065: 'opp_edge': undeclared identifier
D:\a\1\s\src\thor\dijkstras.cc(199): error C2065: 't2': undeclared identifier
D:\a\1\s\src\thor\dijkstras.cc(199): error C2065: 'oppedgeid': undeclared identifier
D:\a\1\s\src\thor\dijkstras.cc(197): error C2789: 'allowed': an object of const-qualified type must be initialized
D:\a\1\s\src\thor\dijkstras.cc(197): note: see declaration of 'allowed'
D:\a\1\s\src\thor\dijkstras.cc(218): error C2065: 'opp_edge': undeclared identifier
D:\a\1\s\src\thor\dijkstras.cc(222): error C2065: 'opp_edge': undeclared identifier
D:\a\1\s\src\thor\dijkstras.cc(222): error C2065: 't2': undeclared identifier
D:\a\1\s\src\thor\dijkstras.cc(245): error C2065: 't2': undeclared identifier
D:\a\1\s\src\thor\dijkstras.cc(246): error C2065: 'oppedgeid': undeclared identifier
D:\a\1\s\src\thor\dijkstras.cc(246): error C2065: 't2': undeclared identifier
D:\a\1\s\src\thor\dijkstras.cc(250): warning C4267: 'initializing': conversion from 'size_t' to 'uint32_t', possible loss of data
D:\a\1\s\src\thor\dijkstras.cc(252): error C2065: 'oppedgeid': undeclared identifier
D:\a\1\s\src\thor\dijkstras.cc(259): error C2065: 'opp_edge': undeclared identifier
D:\a\1\s\src\thor\dijkstras.cc(194): error C2044: illegal continue
D:\a\1\s\src\thor\dijkstras.cc(203): error C2044: illegal continue
D:\a\1\s\src\thor\dijkstras.cc(239): error C2044: illegal continue
D:\a\1\s\src\thor\dijkstras.cc(265): error C2059: syntax error: 'if'
D:\a\1\s\src\thor\dijkstras.cc(265): error C2143: syntax error: missing ';' before '{'
D:\a\1\s\src\thor\dijkstras.cc(265): error C2447: '{': missing function header (old-style formal list?)
D:\a\1\s\src\thor\dijkstras.cc(272): error C2059: syntax error: '}'
D:\a\1\s\src\thor\dijkstras.cc(272): error C2143: syntax error: missing ';' before '}'
D:\a\1\s\src\thor\dijkstras.cc(119): error C2143: syntax error: missing ';' before '{'
D:\a\1\s\src\thor\dijkstras.cc(119): error C2447: '{': missing function header (old-style formal list?)
```