@kevinkreiser 2021-03-30
also a pleasant surprise that the standard deviation of reach is tightend